// api/model/bmfAdmin.session.ts
export class BMFAdminSession {

    userName: string;
    bpNumber: string;
    empSapId: string;
    loggedInSapSecurityRole: string;
    loggedInSapStatus: string;
    loggedInSapUserName: string;
    loggedInSapFirstName: string;
    loggedInSapLastName: string;
    loggedInSapEmailId: string;
    loggedInUserCompCode: string;
    mrktAgrmStatus: string;
    userSapIdInfo: string;
    isUserValid: boolean;
}